+++ SBVUnitTest/SBVUnitTestBuildTime.hs
@@ -2,4 +2,4 @@
 module SBVUnitTestBuildTime (buildTime) where
 
 buildTime :: String
-buildTime = "Sat Nov  7 18:21:16 PST 2015"
+buildTime = "Sun Nov  8 01:01:59 PST 2015"